requirements.txt
requests
beautifulsoup4
python-dateutil
unidecode
pymongo
openpyxl
xlsxwriter
-e .